refactor: isolate passive model loading policy
